+++ src/main/java/ru/skyfire/zeta/dailyrewards/commands/CmdDailyClear.java
@@ -6,23 +6,14 @@ import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.text.Text;
 import ru.skyfire.zeta.dailyrewards.DailyRewards;
-
-import java.io.IOException;
+import ru.skyfire.zeta.dailyrewards.util.CmdUtil;
+import ru.skyfire.zeta.dailyrewards.util.TimeUtil;
 
 public class CmdDailyClear implements CommandExecutor {
     public CommandResult execute(CommandSource sender, CommandContext args) {
-        DailyRewards.getInst().getRootTimeNode().getNode("time").setValue(System.currentTimeMillis());
-        try {
-            DailyRewards.getInst().getTimeConfigLoader().save(DailyRewards.getInst().getRootTimeNode());
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        if (DailyRewards.getInst().hardMode){
-            DailyRewards.getInst().getSqlite().clearDaysHard();
-        }
-        DailyRewards.getInst().getSqlite().clearStatuses();
+        CmdUtil.setNewDay();
 
-        sender.sendMessage(Text.of("Настал новый день, братья!"));
+        sender.sendMessage(Text.of("Every player now can take reward again! Time set to current!"));
         return CommandResult.success();
     }
 }
