+++ src/main/java/ru/skyfire/zeta/dailyrewards/commands/CmdDailyReload.java
@@ -6,8 +6,9 @@ import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.text.Text;
 import ru.skyfire.zeta.dailyrewards.DailyRewards;
-import ru.skyfire.zeta.dailyrewards.Util;
 import ru.skyfire.zeta.dailyrewards.serializers.RewardDeserializer;
+
+import static ru.skyfire.zeta.dailyrewards.util.TextUtil.trans;
 
 public class CmdDailyReload implements CommandExecutor {
     @Override
@@ -16,7 +17,7 @@ public class CmdDailyReload implements CommandExecutor {
         DailyRewards.getInst().initTranslaitionConfig();
         DailyRewards.getInst().initTimeConfig();
         DailyRewards.getInst().setRewardDeserializer(new RewardDeserializer(DailyRewards.getInst().getRootDefNode()));
-        sender.sendMessage(Text.of(Util.trans("command-reload-success")));
+        sender.sendMessage(trans("command-reload-success"));
         return CommandResult.success();
     }
 }
