+++ src/main/java/ru/skyfire/zeta/dailyrewards/commands/CmdDailySet.java
@@ -4,11 +4,6 @@ import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.spec.CommandExecutor;
-import org.spongepowered.api.entity.living.player.Player;
-import org.spongepowered.api.text.Text;
-import ru.skyfire.zeta.dailyrewards.DailyRewards;
-import ru.skyfire.zeta.dailyrewards.Util;
-import ru.skyfire.zeta.dailyrewards.database.SqliteEntry;
 
 public class CmdDailySet implements CommandExecutor {
     @Override
