+++ src/main/java/ru/skyfire/zeta/dailyrewards/commands/CmdDailySetDay.java
@@ -7,8 +7,9 @@ import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.text.Text;
 import ru.skyfire.zeta.dailyrewards.DailyRewards;
-import ru.skyfire.zeta.dailyrewards.Util;
 import ru.skyfire.zeta.dailyrewards.database.SqliteEntry;
+
+import static ru.skyfire.zeta.dailyrewards.util.TextUtil.trans;
 
 public class CmdDailySetDay implements CommandExecutor {
     @Override
@@ -17,19 +18,22 @@ public class CmdDailySetDay implements CommandExecutor {
         Integer targetAmount = args.<Integer>getOne("amount").orElse(null);
 
         if (targetAmount==null || targetPlayer==null){
-            sender.sendMessage(Text.of(Util.trans("command-set-error")));
+            sender.sendMessage(trans("command-set-error"));
             return CommandResult.success();
         }
 
         SqliteEntry sqlite = DailyRewards.getInst().getSqlite();
 
         if(sqlite.getStatus(targetPlayer.getUniqueId())==-1){
-            sender.sendMessage(Text.of(Util.trans("command-set-noplayer")));
+            sender.sendMessage(trans("command-set-noplayer"));
             return CommandResult.success();
         }
 
         sqlite.updateEntry(targetPlayer.getUniqueId(), targetAmount, sqlite.getStatus(targetPlayer.getUniqueId()));
-        sender.sendMessage(Text.of(Util.trans("command-set-day-success")+targetAmount));
+        sender.sendMessage(trans("command-set-day-success")
+                .toBuilder()
+                .append(Text.of(" "+targetAmount))
+                .build());
         return CommandResult.success();
     }
 }
