+++ src/main/java/ru/skyfire/zeta/dailyrewards/commands/CmdDailySetStatus.java
@@ -7,8 +7,9 @@ import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.text.Text;
 import ru.skyfire.zeta.dailyrewards.DailyRewards;
-import ru.skyfire.zeta.dailyrewards.Util;
 import ru.skyfire.zeta.dailyrewards.database.SqliteEntry;
+
+import static ru.skyfire.zeta.dailyrewards.util.TextUtil.trans;
 
 public class CmdDailySetStatus implements CommandExecutor {
     @Override
@@ -17,14 +18,14 @@ public class CmdDailySetStatus implements CommandExecutor {
         Boolean targetStatus = args.<Boolean>getOne("status").orElse(null);
 
         if (targetStatus==null || targetPlayer==null){
-            sender.sendMessage(Text.of(Util.trans("command-set-error")));
+            sender.sendMessage(trans("command-set-error"));
             return CommandResult.success();
         }
 
         SqliteEntry sqlite = DailyRewards.getInst().getSqlite();
 
         if(sqlite.getStatus(targetPlayer.getUniqueId())==-1){
-            sender.sendMessage(Text.of(Util.trans("command-set-noplayer")));
+            sender.sendMessage(trans("command-set-noplayer"));
             return CommandResult.success();
         }
         int status=0;
@@ -33,7 +34,10 @@ public class CmdDailySetStatus implements CommandExecutor {
         }
 
         sqlite.updateEntry(targetPlayer.getUniqueId(), sqlite.getCurrentDay(targetPlayer.getUniqueId()), status);
-        sender.sendMessage(Text.of(Util.trans("command-set-status-success")+status));
+        sender.sendMessage(trans("command-set-status-success")
+                .toBuilder()
+                .append(Text.of(" "+status))
+                .build());
         return CommandResult.success();
     }
 }
