+++ src/main/java/ru/skyfire/zeta/dailyrewards/reward/CmdReward.java
@@ -2,7 +2,8 @@ package ru.skyfire.zeta.dailyrewards.reward;
 
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.entity.living.player.Player;
-import ru.skyfire.zeta.dailyrewards.Util;
+
+import static ru.skyfire.zeta.dailyrewards.util.TextUtil.cmdParser;
 
 public class CmdReward extends Reward {
     String cmd;
@@ -26,9 +27,6 @@ public class CmdReward extends Reward {
 
     @Override
     public void apply(Player player) {
-        Sponge.getCommandManager().process(Sponge.getServer().getConsole(), Util.cmdParser(this.cmd, player));
+        Sponge.getCommandManager().process(Sponge.getServer().getConsole(), cmdParser(this.cmd, player));
     }
 }
-
-//fixme
-//cannot parse command - / in console
