+++ src/main/resources/assets/dailyrewards/DailyRewards.conf
@@ -8,6 +8,8 @@
   messages-amount: 3 // amount of messages which player will recieve if he didn't take his reward. (0 to turn off)
   time-interval: 600 //time interval in seconds between messages. (0 to turn off)
 
+  button: "minecraft:end_crystal"
+
   days{
     1: { //day number - if you will create 1,2 and 4 days, player shall not take reward for 3 day, but he will on 4 day
       reward-message: "&5Wow! Your reaward is diamonds and coal! Yummy!" //Shows on reward take
